chatbot_module/IT HELP DESK  2/requirements.txt
# requirements.txt
openai
python-dotenv
flask
flask-cors
pyjwt
werkzeug
sentence-transformers
chromadb
tiktoken
